+++ src/main/java/uk/gov/ons/ctp/response/casesvc/CaseSvcApplication.java
@@ -295,16 +295,7 @@ public class CaseSvcApplication {
     return new DirectChannel();
   }
 
-  @Bean
-  @ServiceActivator(inputChannel = "pubsubInputChannel")
-  public MessageHandler messageReceiver() {
-    return message -> {
-      log.info("Message arrived! Payload: " + new String((byte[]) message.getPayload()));
-      BasicAcknowledgeablePubsubMessage originalMessage =
-              message.getHeaders().get(GcpPubSubHeaders.ORIGINAL_MESSAGE, BasicAcknowledgeablePubsubMessage.class);
-      originalMessage.ack();
-    };
-  }
+
 
   @Bean
   public PubSubInboundChannelAdapter messageChannelAdapter(
