+++ src/main/java/uk/gov/ons/ctp/response/casesvc/message/ReceiptPubsubReceiver.java
@@ -1,0 +1,25 @@
+package uk.gov.ons.ctp.response.casesvc.message;
+
+import com.godaddy.logging.Logger;
+import com.godaddy.logging.LoggerFactory;
+import org.springframework.cloud.gcp.pubsub.support.BasicAcknowledgeablePubsubMessage;
+import org.springframework.cloud.gcp.pubsub.support.GcpPubSubHeaders;
+import org.springframework.context.annotation.Bean;
+import org.springframework.integration.annotation.ServiceActivator;
+import org.springframework.messaging.MessageHandler;
+
+public class ReceiptPubsubReceiver {
+
+    private static final Logger log = LoggerFactory.getLogger(ReceiptPubsubReceiver.class);
+
+    @Bean
+    @ServiceActivator(inputChannel = "pubsubInputChannel")
+    public MessageHandler messageReceiver() {
+        return message -> {
+            log.info("Message arrived! Payload: " + new String((byte[]) message.getPayload()));
+            BasicAcknowledgeablePubsubMessage originalMessage =
+                    message.getHeaders().get(GcpPubSubHeaders.ORIGINAL_MESSAGE, BasicAcknowledgeablePubsubMessage.class);
+            originalMessage.ack();
+        };
+    }
+}
